+++ app/src/main/java/com/chillibits/colorconverter/shared/Constants.kt
@@ -2,7 +2,7 @@
  * Copyright © Marc Auberer 2020. All rights reserved
  */
 
-package com.chillibits.colorconverter.tools
+package com.chillibits.colorconverter.shared
 
 object Constants {
     // Request codes
