+++ app/src/main/java/com/chillibits/colorconverter/shared/Extensions.kt
@@ -2,7 +2,7 @@
  * Copyright © Marc Auberer 2020. All rights reserved
  */
 
-package com.chillibits.colorconverter.tools
+package com.chillibits.colorconverter.shared
 
 import android.content.ClipData
 import android.content.ClipboardManager
