+++ app/src/main/java/com/chillibits/colorconverter/tools/ColorName.kt
@@ -4,8 +4,12 @@
 
 package com.chillibits.colorconverter.tools
 
-class ColorName(private var name: String, private var r: Int, private var g: Int, private var b: Int) {
+class ColorName(
+    var name: String,
+    private var r: Int,
+    private var g: Int,
+    private var b: Int
+) {
     fun computeMSE(pixR: Int, pixG: Int, pixB: Int) =
         ((pixR - r) * (pixR - r) + (pixG - g) * (pixG - g) + (pixB - b) * (pixB - b)) / 3
-    fun getName(): String = name
 }
