+++ app/src/main/res/values-de/strings_colors.xml
@@ -9,7 +9,7 @@
  Release: Working copy
  Locale: de, German
  Exported by: ChilliBits
- Exported at: Sun, 02 Aug 2020 00:22:26 +0100
+ Exported at: Fri, 04 Sep 2020 00:45:25 +0100
 -->
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <string name="color_alice_blue">Aliceblau</string>
