+++ app/src/main/res/values-fr/strings.xml
@@ -9,7 +9,7 @@
  Release: Working copy
  Locale: fr, French
  Exported by: ChilliBits
- Exported at: Sun, 02 Aug 2020 00:22:26 +0100
+ Exported at: Fri, 04 Sep 2020 00:45:25 +0100
 -->
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <string name="ok">OK</string>
